+++ src/libFLAC/libFLAC_dynamic.vcxproj.filters
@@ -131,7 +131,7 @@
     <ClInclude Include="..\..\include\share\safe_str.h">
       <Filter>Public Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\include\share\windows_unicode_filenames.h">
+    <ClInclude Include="..\..\include\share\win_utf8_io.h">
       <Filter>Public Header Files</Filter>
     </ClInclude>
   </ItemGroup>
@@ -223,7 +223,7 @@
     <ClCompile Include="window.c">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="windows_unicode_filenames.c">
+    <ClCompile Include="..\share\win_utf8_io\win_utf8_io.c">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
