+++ src/libFLAC/libFLAC_static.vcproj
@@ -321,7 +321,7 @@
 				>
 			</File>
 			<File
-				RelativePath="..\..\include\share\windows_unicode_filenames.h"
+				RelativePath="..\..\include\share\win_utf8_io.h"
 				>
 			</File>
 		</Filter>
@@ -447,7 +447,7 @@
 				>
 			</File>
 			<File
-				RelativePath=".\windows_unicode_filenames.c"
+				RelativePath="..\share\win_utf8_io\win_utf8_io.c"
 				>
 			</File>
 		</Filter>
