+++ src/test_grabbag/cuesheet/Makefile.lite
@@ -32,7 +32,11 @@ INCLUDES = -I./include -I$(topdir)/include
 ifeq ($(OS),Darwin)
     EXPLICIT_LIBS = $(libdir)/libgrabbag.a $(libdir)/libreplaygain_analysis.a $(libdir)/libFLAC.a $(OGG_EXPLICIT_LIBS) -lm
 else
-    LIBS = -lgrabbag -lreplaygain_analysis -lFLAC $(OGG_LIBS) -lm
+ifeq ($(findstring Windows,$(OS)),Windows)
+    LIBS = -lgrabbag -lreplaygain_analysis -lFLAC -lwin_utf8_io $(OGG_LIBS) -lm
+else
+    LIBS = -lgrabbag -lreplaygain_analysis -lFLAC -lgrabbag $(OGG_LIBS) -lm
+endif
 endif
 
 SRCS_C = \
