+++ src/test_seeking/Makefile.lite
@@ -32,7 +32,11 @@ INCLUDES = -I../libFLAC/include -I$(topdir)/include
 ifeq ($(OS),Darwin)
     EXPLICIT_LIBS = $(libdir)/libFLAC.a $(OGG_EXPLICIT_LIBS) -lm
 else
+ifeq ($(findstring Windows,$(OS)),Windows)
+    LIBS = -lFLAC -lwin_utf8_io $(OGG_LIBS) -lm
+else
     LIBS = -lFLAC $(OGG_LIBS) -lm
+endif
 endif
 
 SRCS_C = \
